Check for syntax errors using RubyVM
